pyproject.toml
[project]
name = "vs-code-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aspose-cells-python>=25.1.0",
    "duckdb>=1.1.3",
    "ipykernel>=6.29.5",
    "polars>=1.21.0",
    "pyxlsb>=1.0.10",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.3",
]

# helps pytest find the source code
[tool.pytest.ini_options]
pythonpath = "src"
